+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - bumped types dependency for node
 - bumped all dev dependencies - includes some prettier changes
 - package-lock updated to version 2
+- fix #378 exit code not set on parse failure. A proper error will be set on the stream now.
 
 ## 7.0.0